\name{nlqmm-internal}
\title{Internal nlqmm objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{bootData}
\alias{hfun_nlqmm}
\alias{C_hfun_nlqmm}
\alias{C_loss_nlqmm}
\alias{loglik_nlqmm}
\alias{parsef}
\alias{modalRe_nlqmm}
\description{Internal nlqmm objects.}
\details{These are not to be called by the user.}
\keyword{internal}