\name{nlqmm-package}
\alias{nlqmm-package}
\docType{package}
\title{
Nonlinear Quantile Mixed Models}

\description{
Fit nonlinear quantile mixed models
}

\details{
\tabular{ll}{
Package: \tab nlqmm\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2020-04-15\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}

\references{
Geraci M (2019). Modelling and estimation of nonlinear quantile regression with clustered data. Computational Statistics and Data Analysis, 136, 30-46.
}

\keyword{quantile regression}
