\name{print.nlqmm}
\alias{print.nlqmm}

\title{
Print an \code{nlqmm} Object
}
\description{
Print an object generated by \code{\link{nlqmm}}.
}
\usage{
\method{print}{nlqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{nlqmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{nlqmm}}
}

\keyword{print}
