\name{nlqmm-package}
\alias{nlqmm-package}
\docType{package}
\title{
Nonlinear Quantile Mixed Models}

\description{
Fit nonlinear quantile mixed models
}

\details{
\tabular{ll}{
Package: \tab nlqmm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2019-02-25\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}

\references{
Geraci M (2018). Modelling and estimation of nonlinear quantile regression with clustered data. Computational Statistics and Data Analysis, doi:10.1016/j.csda.2018.12.005. Available at \url{https://arxiv.org/abs/1712.09981}.
}

\keyword{quantile regression}
